BProbeMod now skips determination of Global Minimum if a starting point is specified with in the Initializtion.
BProbeMod now skips determination of Global Minimum if a starting point is specified with in the Initializtion. This cuts out a large chunk of expensive computation if the dimensions of the matrix set are large. This means that for now one cannot use the minimal energy point as a reset place without recompiling the operator unless one did NOT specifiy a starting point.